Update Download method and delete some dead code
Download is now the method of type File.
No need to lock map.
We just change the value of passed File, which at time of caliing func is already a member of slice []*File.
Also all declarations of values are in right order now, just before actual using of them in code.